Fetch on demand.
Regional forecast is fetched "on demand", ie when the user
selects the Regional Forecast tab in the skin. (Requires
corresponding skin update). This lowers the load and response
time, only fetching and serving data that the user needs.

A timestamp is added to the data which is checked to see if
an update is needed.